Complete dynamic sensor placement system, no debug
In this version I remove sensorAttachmentType from
ArrowbotParameters and transfer all of its processing back to
main.cpp. For that I add two new methods to ArrowbotSimulator
class, setSensorAttachmentMatrix(matrix) and
placeSensors(NumericVector). The former sets the attachemnt matrix
directly and is used if the attachment is static; the latter is
used in variable sensor attachemnt implementation.

Aside from that, the code for handling variable sensor attachment
has been added to main.cpp and several queries to the old
Individual/Phenotype interface were found und updated.

Not a thing of all that is debugged yet.

	modified:   arrowbotSimulator.cpp
	modified:   arrowbotSimulator.h
	modified:   evclib
	modified:   main.cpp